samples: keys: Check and write lifecycle state
The identity key generation sample now checks the
PSA lifecycle state of the device before it writes
the hardware unique keys and the initial attestation
keys to the device. It also updates the PSA lifecycle
state of the device to PROVISIONING. This will
allow the provisioning of the TF-M in a later stage.

Ref: NCSDK-17001

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>
Signed-off-by: Markus Swarowsky <markus.swarowsky@nordicsemi.no>